wms/requirements.txt
# WMS 산업용 협동로봇 전문 시스템 필수 패키지
# 크롤링 및 데이터 수집
arxiv==2.1.0
scholarly==1.7.11
requests==2.31.0
beautifulsoup4==4.12.2

# 텍스트 처리 및 분석
pandas==2.1.4
numpy==1.24.3
nltk==3.8.1
scikit-learn==1.3.2

# PDF 처리 (다중 백업)
PyPDF2==3.0.1
PyMuPDF==1.23.14
pdfplumber==0.10.3

# 벡터 데이터베이스
chromadb==0.4.22
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# 웹 API 프레임워크
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-dotenv==1.0.0
loguru==0.7.2

# 기타 유틸리티
tqdm==4.66.1
pathlib2==2.3.7